backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
joblib==1.4.2
pandas==2.2.2
scikit-learn==1.4.2
python-dotenv==1.0.1